+++ wutengine/src/builtins/components/mod.rs
@@ -1,10 +1,12 @@
 mod camera;
+mod input;
 mod material;
 mod mesh;
 mod name;
 mod transform;
 
 pub use camera::*;
+pub use input::*;
 pub use material::*;
 pub use mesh::*;
 pub use name::*;