+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This file documents the most important changes for each released version.
 - Initial support for optional components in ECS queries
 - Camera component now influences rendering through its position (as determined by its Transform component) and its perspective settings
 - Engine plugins can now listen and respond to raw window and device events
+- Basic keyboard input handling
 
 ## [v0.0.6]
 - Complete refactor of core ECS system. We now use archetypes. It was a lot of work.